Patched bug where EZ scheduler jobs were not re-initiating upon server reboot

+++ server/server_util_funcs/jobfunction.js
@@ -3,15 +3,16 @@ const { red } = require("./red");
 
 const jobFunction = async (crontype, macAddress, oneTime, unifi, prisma) => { // for crons
     try {
+        if (crontype === undefined) throw new Error("**block/allow must be specified for jobFunction, undefined was passed!**");
         if (crontype === 'allow') {
             console.log('unifi === undefined \t', unifi === undefined);
-            const confirmAllow = await unifi.unblockClient(macAddress);
+            const confirmAllow = await unifi?.unblockClient(macAddress);
             console.log(`${macAddress} has been unblocked: ${confirmAllow}`);
             if (oneTime) {
                 deleteCompletedJobs(prisma);
             }
         } else if (crontype === 'block') {
-            const confirmBlocked = await unifi.blockClient(macAddress);
+            const confirmBlocked = await unifi?.blockClient(macAddress);
             console.log(`${macAddress} has been blocked: ${confirmBlocked}`);
             if (oneTime) {
                 deleteCompletedJobs(prisma);
